# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API backend de la plateforme RAYCARD (wallet Mobile Money + cartes Visa virtuelles).",
        "title": "RAYCARD API",
        "contact": {
            "name": "RAYCARD — HR-Skills SARL"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/backoffice/cartes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retourne les cartes tous utilisateurs confondus, filtrables par utilisateur et par statut.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "\"2. Admin - Carte\""
                ],
                "summary": "Liste des cartes (back-office)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filtre par utilisateur",
                        "name": "utilisateur_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtre par statut (active, gelee, annulee)",
                        "name": "statut",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/carte.CarteDTO"
                            }
                        }
                    },
                    "401": {
                        "description": "non authentifié",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "403": {
                        "description": "réservé aux administrateurs",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "500": {
                        "description": "erreur interne",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    }
                }
            }
        },
        "/backoffice/cartes/{id}/annuler": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Détruit définitivement n'importe quelle carte active ou gelée et rembourse au wallet ce qu'il restait dessus. Tracé dans l'audit log.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "\"2. Admin - Carte\""
                ],
                "summary": "Annulation d'une carte (back-office)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la carte",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/carte.CarteDTO"
                        }
                    },
                    "401": {
                        "description": "non authentifié",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "403": {
                        "description": "réservé aux administrateurs",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "404": {
                        "description": "carte ou wallet introuvable",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "409": {
                        "description": "une opération wallet est déjà en cours",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "422": {
                        "description": "carte déjà annulée, ou wallet gelé",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "500": {
                        "description": "erreur interne",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    }
                }
            }
        },
        "/backoffice/cartes/{id}/degel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Réactive n'importe quelle carte gelée, sans vérification de propriétaire. Tracé dans l'audit log.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "\"2. Admin - Carte\""
                ],
                "summary": "Dégel d'une carte (back-office)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la carte",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/carte.CarteDTO"
                        }
                    },
                    "401": {
                        "description": "non authentifié",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "403": {
                        "description": "réservé aux administrateurs",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "404": {
                        "description": "carte introuvable",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "422": {
                        "description": "la carte n'est pas gelée",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "500": {
                        "description": "erreur interne",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    }
                }
            }
        },
        "/backoffice/cartes/{id}/gel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Bloque n'importe quelle carte active, sans vérification de propriétaire. Tracé dans l'audit log.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "\"2. Admin - Carte\""
                ],
                "summary": "Gel d'une carte (back-office)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la carte",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/carte.CarteDTO"
                        }
                    },
                    "401": {
                        "description": "non authentifié",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "403": {
                        "description": "réservé aux administrateurs",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "404": {
                        "description": "carte introuvable",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "422": {
                        "description": "la carte n'est pas active",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    },
                    "500": {
                        "description": "erreur interne",
                        "schema": {
                            "$ref": "#/definitions/commun.ErreurDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "carte.CarteDTO": {
            "type": "object",
            "properties": {
                "devise": {
                    "type": "string",
                    "example": "XOF"
                },
                "id": {
                    "type": "string",
                    "example": "3fa2c1e4-9b5d-4a2e-8c1a-0e2f6a7b8c9d"
                },
                "label": {
                    "type": "string",
                    "example": "Carte courses"
                },
                "montant_charge_centimes": {
                    "type": "integer",
                    "example": 10000
                },
                "solde_centimes": {
                    "description": "SoldeCentimes est le dernier solde connu, mis à jour périodiquement\n(pas en temps réel — voir carte.Carte.SoldeCentimes).",
                    "type": "integer",
                    "example": 7500
                },
                "statut": {
                    "type": "string",
                    "example": "active"
                }
            }
        },
...
        "commun.ErreurDTO": {
            "type": "object",
            "properties": {
                "erreur": {
                    "type": "string",
                    "example": "un compte existe déjà avec cet email"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}